requirements.prod.txt
fastapi
uvicorn
boto3
python-dotenv
onnxruntime
numpy
pydantic